refactor: rename `autocommands.lua` to `autocmds.lua`
