+++ lua/meowim/config/autocmds.lua
